base refactor
