+++ mmocr/models/common/__init__.py
@@ -1,7 +1,9 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from . import backbones, losses
+from . import backbones, layers, losses, modules
 
 from .backbones import *  # NOQA
 from .losses import *  # NOQA
+from .layers import *  # NOQA
+from .modules import *  # NOQA
 
-__all__ = backbones.__all__ + losses.__all__
+__all__ = backbones.__all__ + losses.__all__ + layers.__all__ + modules.__all__
