+++ mmocr/models/common/layers/__init__.py
@@ -1,0 +1,6 @@
+from .transformer_layers import (TFDecoderLayer, TFEncoderLayer, get_pad_mask,
+                                 get_subsequent_mask)
+
+__all__ = [
+    'TFEncoderLayer', 'TFDecoderLayer', 'get_pad_mask', 'get_subsequent_mask'
+]
