+++ mmocr/models/common/modules/__init__.py
@@ -1,0 +1,8 @@
+from .transformer_module import (MultiHeadAttention, PositionalEncoding,
+                                 PositionwiseFeedForward,
+                                 ScaledDotProductAttention)
+
+__all__ = [
+    'ScaledDotProductAttention', 'MultiHeadAttention',
+    'PositionwiseFeedForward', 'PositionalEncoding'
+]
