+++ mmocr/models/textrecog/encoders/__init__.py
@@ -1,11 +1,11 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .base_encoder import BaseEncoder
 from .channel_reduction_encoder import ChannelReductionEncoder
+from .nrtr_encoder import NRTREncoder
 from .sar_encoder import SAREncoder
 from .satrn_encoder import SatrnEncoder
-from .transformer_encoder import TFEncoder
 
 __all__ = [
-    'SAREncoder', 'TFEncoder', 'BaseEncoder', 'ChannelReductionEncoder',
+    'SAREncoder', 'NRTREncoder', 'BaseEncoder', 'ChannelReductionEncoder',
     'SatrnEncoder'
 ]
