+++ mmocr/models/textrecog/layers/__init__.py
@@ -4,17 +4,10 @@ from .dot_product_attention_layer import DotProductAttentionLayer
 from .lstm_layer import BidirectionalLSTM
 from .position_aware_layer import PositionAwareLayer
 from .robust_scanner_fusion_layer import RobustScannerFusionLayer
-from .transformer_layer import (Adaptive2DPositionalEncoding,
-                                MultiHeadAttention, PositionalEncoding,
-                                PositionwiseFeedForward, SatrnEncoderLayer,
-                                TransformerDecoderLayer,
-                                TransformerEncoderLayer, get_pad_mask,
-                                get_subsequent_mask)
+from .satrn_layers import Adaptive2DPositionalEncoding, SatrnEncoderLayer
 
 __all__ = [
-    'BidirectionalLSTM', 'MultiHeadAttention', 'PositionalEncoding',
-    'Adaptive2DPositionalEncoding', 'PositionwiseFeedForward', 'BasicBlock',
+    'BidirectionalLSTM', 'Adaptive2DPositionalEncoding', 'BasicBlock',
     'Bottleneck', 'RobustScannerFusionLayer', 'DotProductAttentionLayer',
-    'PositionAwareLayer', 'get_pad_mask', 'get_subsequent_mask',
-    'TransformerDecoderLayer', 'TransformerEncoderLayer', 'SatrnEncoderLayer'
+    'PositionAwareLayer', 'SatrnEncoderLayer'
 ]
